Main Source Code with Database / CSS Layout
This is the main source code for the project that I have added that has the models, views, and controller used for this project. The layouts I've added were tweaked a bit with very minor css changes. This also includes validation checking for user input so that the program doesn't crash.